Clarify type and contents for payload
Remove confusing duplication of referencing JSON, XML and TEXT, the caps imply the "type" rather than the contents of the string. 

Given the contents can be text not really sure what the point of the saying it has to be well formed json, xml or text. If that is, because the type is checked and the content-type set according to what is determined. The it should say so